ai-service/requirements.txt

fastapi
uvicorn[standard]
python-multipart
pillow
httpx
requests
qdrant-client
sentence-transformers
psycopg2-binary
python-dotenv